requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
channels==3.0.4
charset-normalizer==2.0.4
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.4.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.5
django-heroku==0.3.1
django-import-export==2.5.0
django-multiselectfield==0.1.12
et-xmlfile==1.1.0
fonttools==4.27.1
future==0.18.2
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
joblib==1.0.1
MarkupPy==1.14
msgpack==1.0.2
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydyf==0.1.1
pyOpenSSL==20.0.1
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
random2==1.0.1
reportlab==3.6.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
tablib==3.0.0
threadpoolctl==3.0.0
tinycss2==1.1.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
zopfli==0.1.8
